back-end/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
PyPDF2

